Fixed issue with windows paths.
Removed a JSON.parse() call who's purpose was to remove quotes from strings, and replaced it with substr.

There was a check to see if an include path was referencing the root directory. The check was looking for the first character to be '/', which is not how Windows works. So, replaced that with a regex test.